Conflicts:
	Airship/Common/UAConfig.h
	Airship/Common/UAConfig.m
	AirshipLib/AirshipLib.xcodeproj/project.pbxproj
	AirshipLib/Config.xcconfig
	InboxSample/InboxSample.xcodeproj/project.pbxproj
	PushSample/PushSample.xcodeproj/project.pbxproj

+++ Airship/Inbox/UAInboxMessageList.h
@@ -199,7 +199,7 @@ typedef enum {
  * The number of messages that are currently unread or -1
  * if the message list is not loaded.
  */
-@property(assign) int unreadCount;
+@property(assign) NSUInteger unreadCount;
 
 /**
  * YES if a batch update is currently being performed on the message list,
